fix discovery middleware
