+++ packages/suite/src/actions/suite/suiteActions.ts
@@ -153,7 +153,6 @@ export const selectDevice = (device?: Device | TrezorDevice) => async (
 ) => {
     // 1. check if device is not locked by device request or application itself (for example onboarding process)
     const { locks } = getState().suite;
-    if (locks.includes(SUITE.LOCK_TYPE.DEVICE)) return;
     // 2. check if device is acquired
     if (device && device.features && !locks.includes(SUITE.LOCK_TYPE.ROUTER)) {
         // 3. device is not initialized, redirect to "onboarding"
